requirements.txt
snowflake-connector-python==2.3.10
snowflake-sqlalchemy==1.2.4
seaborn==0.11.1
pandas==1.2.3
numpy==1.20.1
SQLAlchemy==1.3.23
alembic==1.5.7
SQLAlchemy-Utils==0.37.6
psycopg2==2.8.6